Don't compare unsigned int with signed
Error: COMPILER_WARNING:
woff2-1.0.2/src/font.cc: scope_hint: In function 'int woff2::NumGlyphs(const woff2::Font&)'
woff2-1.0.2/src/font.cc:330:26: warning: comparison of integer expressions of different signedness: 'const uint32_t' {aka 'const unsigned int'} and 'int' [-Wsign-compare]
   if (loca_table->length < loca_record_size) {
       ~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~
  328|     int index_fmt = IndexFormat(font);
  329|     int loca_record_size = (index_fmt == 0 ? 2 : 4);
  330|->   if (loca_table->length < loca_record_size) {
  331|       return 0;
  332|     }

Error: COMPILER_WARNING:
woff2-1.0.2/src/normalize.cc:15: included_from: Included from here.
woff2-1.0.2/src/normalize.cc: scope_hint: In function 'bool woff2::{anonymous}::MakeEditableBuffer(woff2::Font*, int)'
woff2-1.0.2/src/normalize.cc:100:24: warning: comparison of integer expressions of different signedness: 'int' and 'uint32_t' {aka 'unsigned int'} [-Wsign-compare]
   if (PREDICT_FALSE(sz > table->length)) {
                     ~~~^~~~~~~~~~~~~~~
woff2-1.0.2/src/port.h:48:44: note: in definition of macro 'PREDICT_FALSE'
 #define PREDICT_FALSE(x) (__builtin_expect(x, 0))
                                            ^
   98|     uint8_t* buf = &table->buffer[0];
   99|     memcpy(buf, table->data, table->length);
  100|->   if (PREDICT_FALSE(sz > table->length)) {
  101|       memset(buf + table->length, 0, sz - table->length);
  102|     }

Error: COMPILER_WARNING:
woff2-1.0.2/src/woff2_info.cc: scope_hint: In function 'int main(int, char**)'
woff2-1.0.2/src/woff2_info.cc:125:24: warning: comparison of integer expressions of different signedness: 'int' and 'uint32_t' {aka 'unsigned int'} [-Wsign-compare]
     for (auto i = 0; i < numFonts; i++) {
                      ~~^~~~~~~~~~
  123|       printf("CollectionHeader 0x%08x %d fonts\n", version, numFonts);
  124|
  125|->     for (auto i = 0; i < numFonts; i++) {
  126|         uint32_t numTables, flavor;
  127|         if (!woff2::Read255UShort(&file, &numTables)) return 1;

Error: COMPILER_WARNING:
woff2-1.0.2/src/woff2_info.cc:131:26: warning: comparison of integer expressions of different signedness: 'int' and 'uint32_t' {aka 'unsigned int'} [-Wsign-compare]
       for (auto j = 0; j < numTables; j++) {
                        ~~^~~~~~~~~~~
  129|         printf("CollectionFontEntry %d flavor 0x%08x %d tables\n", i, flavor,
  130|             numTables);
  131|->       for (auto j = 0; j < numTables; j++) {
  132|           uint32_t table_idx;
  133|           if (!woff2::Read255UShort(&file, &table_idx)) return 1;

+++ src/normalize.cc
@@ -97,7 +97,7 @@ bool MakeEditableBuffer(Font* font, int tableTag) {
   table->buffer.resize(sz);
   uint8_t* buf = &table->buffer[0];
   memcpy(buf, table->data, table->length);
-  if (PREDICT_FALSE(sz > table->length)) {
+  if (PREDICT_FALSE(static_cast<uint32_t>(sz) > table->length)) {
     memset(buf + table->length, 0, sz - table->length);
   }
   table->data = buf;
